Move HREF type alias to link module
