--- pystac/types.py
@@ -1,4 +1,0 @@
-import os
-from typing import AnyStr, Union
-
-HREF = Union[str, os.PathLike]
